feat(http): Compat with non-tokio runtimes (test with async-std)
It seems some of our users sometimes use async-std, however since
we user Reqwest for our HTTP reverse bindings, we now need a tokio
runtime to create a Core object.

If one is present when a Core is created, we take a handle to it.
Otherwise we start our own tokio multithreaded runtime that lives
as long as the Core session.