fix: Add Content-Length: 0 header when sending empty bodies
 This scenario happens specifically when we create a device, stop it,
and restart it from the same storage, but ONLY when we
use a specific TLS configuration! (Either OpenSSL default,
or a manually configured Rustls with only ALPN 1.1 enabled...)

Because all our other tests use an in-memory storage, and because they
use rustls default settings (ALPN 2.0),
they did not hit this case before.