feat: Support system certificate store and SSL_CERT_FILE
On desktops, we turn on the support by default because
the extra dependency is not a problem.
On mobile the system certificate store is not fully supported by
Rustls, and we want to save space, so only enable webpki by default.

However it is still possible to use self-signed certs on mobile
with this feature, by uploading the CA cert to the device and
setting its path in the SSL_CERT_FILE env var.