fix(testapp): Spawn a whole thread for async drop, don't block_on
block_on in cheap and good-enough most of the time,
but is, in fact, exceptionally unwise.
There is just no very good alternative until async drop becomes a
real feature of the language instead of a flaming hot shrug emoji.

Replace it with a whole scoped thread running its own Tokio runtime.
It's a little heavyweight to spawn a whole thread and runtime
just to make one (1) HTTP request, but at least it cannot deadlock
(the parent thread can block without problem, since the runtime
running the HTTP request is on the new thread, not in the parent)