fix: Copy non-libc android-specific NDK libs as armerge inputs
We want libs like libcompiler_rt-extras.a or libunwind.a
to be added to the armerge static lib.
Previously this was handled by making the legacy NDK a dependency
of sdk-native, but with the new NDK recipe that hack is gone,
so we copy explicitely the libs we need from the NDK package.